renamed lvq module to glvq
