+++ sklearn/glvq/glvq.py
